Combined all protractors into one conf
Also using an env variable to choose whether to run sauce connect or browserstack local
