--- tests/protractor-firefox.conf.js
@@ -1,23 +1,0 @@
-var helper = require('./firefox-helper.js');
-
-exports.config = {
-  allScriptsTimeout: 11000,
-  directConnect: true,
-  firefoxPath: process.env.FIREFOX_BIN,
-
-  specs: [
-    'e2e/scenarios.js'
-  ],
-
-  baseUrl: 'http://localhost:5000/',
-
-  getMultiCapabilities: helper.getFirefoxProfile,
-
-  params: {
-    testScreenSharing: true
-  },
-
-  jasmineNodeOpts: {
-    defaultTimeoutInterval: 30000
-  }
-};
