--- tests/protractor-ie11.conf.js
@@ -1,24 +1,0 @@
-exports.config = {
-  allScriptsTimeout: 30000,
-
-  specs: [
-    'e2e/iesmoketest.js'
-  ],
-
-  capabilities: {
-    'browserName': 'internet explorer',
-    'platform': 'ANY',
-    'version': '11',
-    'ignoreZoomSetting': true
-  },
-
-  seleniumAddress: 'http://ie11.dev:4444/wd/hub',
-
-  baseUrl: 'http://adam.local:5000/',
-
-  framework: 'jasmine',
-
-  jasmineNodeOpts: {
-    defaultTimeoutInterval: 60000
-  }
-};
