+++ tests/run-tests
@@ -1,6 +1,7 @@
 #!/bin/bash
 #
-# Installs Chrome or Firefox if missing and runs Karma and Protractor tests
+# Installs Chrome or Firefox if missing, starts Browserstack Local or Sauce Connect if needed
+# and runs Karma and Protractor tests
 #
 BINDIR=./browsers/bin
 
@@ -22,10 +23,18 @@ if [ -z $TRAVIS ]; then
 		firefox-unstable) BROWSERBIN="/Applications/FirefoxDeveloperEdition.app/Contents/MacOS/firefox";;
 	esac
 else
-	BROWSERBIN=$BINDIR/$BROWSER-$BVER
+	# We can't install ie on Travis
+	if [ $BROWSER != 'ie' ]; then
+		BROWSERBIN=$BINDIR/$BROWSER-$BVER
+	fi
 fi
 
-if [ $BROWSER == 'ie' ]; then
+if [ -n "$BROWSERBIN" ] && [ ! -x $BROWSERBIN ]; then
+	echo "Installing browser."
+	./node_modules/travis-multirunner/setup.sh
+fi
+
+if [ -n "$SAUCELABS" ]; then
 	echo "Starting Sauce Connect"
 	if [ ! -x ./sc-4.3.11-*/bin/sc ]; then
 		wget http://saucelabs.com/downloads/sc-4.3.11-linux.tar.gz
@@ -38,32 +47,9 @@ if [ $BROWSER == 'ie' ]; then
 	while [ ! -f $READY_FILE ]; do
 		sleep .5
 	done
+fi
 
-	echo "Starting Karma"
-	karma start tests/karma.conf.js --single-run --browsers "IE$BVER"
-	karma=$?
-
-	echo "Running Protractor tests"
-	protractor tests/protractor-saucelabs-ie.conf.js
-	protractor=$?
-
-	echo "Shutting down Sauce Connect"
-	kill -2 $s_pid
-	wait $s_pid
-else
-	if [ ! -x $BROWSERBIN ]; then
-		echo "Installing browser."
-		./node_modules/travis-multirunner/setup.sh
-	fi
-	PATH=$PATH:./node_modules/.bin
-	# Karma uses ${BROWSER}_BIN, eg. CHROME_BIN, to override the path to the browser
-	BROWSER_UPPER=$(echo "$BROWSER" | tr '[:lower:]' '[:upper:]')
-	BROWSER_CAMEL=${BROWSER_UPPER:0:1}${BROWSER:1}
-	export ${BROWSER_UPPER}_BIN="$BROWSERBIN"
-	echo "Starting Karma for browser $BROWSER_CAMEL using $BROWSERBIN"
-	karma start tests/karma.conf.js --single-run --browsers $BROWSER_CAMEL
-	karma=$?
-
+if [ -n "$BROWSERSTACK" ]; then
 	echo "Starting BrowserStackLocal"
 	if [ ! -x ./BrowserStackLocal ]; then
 		wget http://www.browserstack.com/browserstack-local/BrowserStackLocal-linux-x64.zip
@@ -72,11 +58,28 @@ else
 	./BrowserStackLocal $BROWSERSTACK_KEY localhost,8001,0 > /dev/null &
 	b_pid=$!
 	sleep 10
+fi
 
-	webdriver-manager update
-	echo "Running Protractor tests"
-	protractor tests/protractor-browserstack-$BROWSER.conf.js
-	protractor=$?
+PATH=$PATH:./node_modules/.bin
+# Karma uses ${BROWSER}_BIN, eg. CHROME_BIN, to override the path to the browser
+BROWSER_UPPER=$(echo "$BROWSER" | tr '[:lower:]' '[:upper:]')
+export ${BROWSER_UPPER}_BIN="$BROWSERBIN"
+echo "Starting Karma for browser $BROWSER"
+karma start tests/karma.conf.js --single-run
+karma=$?
+
+webdriver-manager update
+echo "Running Protractor tests"
+protractor tests/protractor.conf.js
+protractor=$?
+
+if [ ! -z $s_pid ]; then
+	echo "Shutting down Sauce Connect"
+	kill -2 $s_pid
+	wait $s_pid
+fi
+
+if [ ! -z $b_pid ]; then
 	echo "Shutting down BrowserstackLocal $b_pid"
 	kill -2 $b_pid
 	wait $b_pid
